refactor(clinics): move clinic availability methods inside clinic service
