+++ src/main/java/com/medspace/application/usecase/clinicAvailability/DeleteClinicAvailabilityByIdUseCase.java
@@ -1,6 +1,5 @@
 package com.medspace.application.usecase.clinicAvailability;
 
-import com.medspace.application.service.ClinicAvailabilityService;
 import com.medspace.application.service.ClinicService;
 import com.medspace.domain.model.ClinicAvailability;
 import io.quarkus.security.ForbiddenException;
@@ -11,20 +10,18 @@ import jakarta.transaction.Transactional;
 @ApplicationScoped
 public class DeleteClinicAvailabilityByIdUseCase {
     @Inject
-    ClinicAvailabilityService clinicAvailabilityService;
-    @Inject
     ClinicService clinicService;
 
     @Transactional
     public void execute(Long availabilityId, Long userId) {
         ClinicAvailability clinicAvailability =
-                clinicAvailabilityService.getAvailabilityById(availabilityId);
+                clinicService.getClinicAvailabilityById(availabilityId);
         Long clinicId = clinicAvailability != null ? clinicAvailability.getClinic().getId() : null;
         Boolean isOwner = clinicService.validateClinicOwnership(clinicId, userId);
         if (!isOwner) {
             throw new ForbiddenException("Delete unauthorized");
         }
 
-        clinicAvailabilityService.deleteAvailabilityById(availabilityId);
+        clinicService.deleteClinicAvailabilityById(availabilityId);
     }
 }
