+++ src/main/java/com/medspace/application/usecase/clinicAvailability/UpdateClinicAvailabilityUseCase.java
@@ -1,6 +1,5 @@
 package com.medspace.application.usecase.clinicAvailability;
 
-import com.medspace.application.service.ClinicAvailabilityService;
 import com.medspace.application.service.ClinicService;
 import com.medspace.domain.model.ClinicAvailability;
 import io.quarkus.security.ForbiddenException;
@@ -11,13 +10,11 @@ import jakarta.transaction.Transactional;
 @ApplicationScoped
 public class UpdateClinicAvailabilityUseCase {
     @Inject
-    ClinicAvailabilityService clinicAvailabilityService;
-    @Inject
     ClinicService clinicService;
 
     @Transactional
     public ClinicAvailability execute(Long id, ClinicAvailability clinicAvailability, Long userId) {
-        ClinicAvailability existingAvailability = clinicAvailabilityService.getAvailabilityById(id);
+        ClinicAvailability existingAvailability = clinicService.getClinicAvailabilityById(id);
         Long clinicId =
                 existingAvailability != null ? existingAvailability.getClinic().getId() : null;
         Boolean isOwner = clinicService.validateClinicOwnership(clinicId, userId);
@@ -25,6 +22,6 @@ public class UpdateClinicAvailabilityUseCase {
             throw new ForbiddenException("Update unauthorized");
         }
 
-        return clinicAvailabilityService.updateAvailabilityById(id, clinicAvailability);
+        return clinicService.updateClinicAvailabilityById(id, clinicAvailability);
     }
 }
